Migrate Qml files to shared
