--- src/RogueCollectionQml/RoguePlugin/key_utility.h
@@ -1,5 +1,0 @@
-#pragma once
-
-bool IsLetterKey(unsigned int keycode);
-bool IsDirectionKey(unsigned int keycode);
-bool IsCapsLockOn();
