+++ src/Shared/Frontend/key_utility.h
@@ -1,4 +1,5 @@
 #pragma once
 
-bool IsLetterKey(unsigned char keycode);
-bool IsDirectionKey(unsigned char keycode);
+bool IsLetterKey(unsigned int keycode);
+bool IsDirectionKey(unsigned int keycode);
+bool IsCapsLockOn();
